refactor(debug_server): extract voice endpoint family (refs #332)
Fourth per-family extract from debug_server.c, stacked on K144 (#336),
OTA (#337), and codec (#338).

Endpoints owned by this PR:
  GET  /voice            — voice pipeline state snapshot
  POST /voice/reconnect  — force voice WS reconnect
  POST /voice/cancel     — cancel in-flight voice turn
  POST /voice/clear      — clear Dragon + Tab5 conversation history

These are the 4 voice-pipeline-control endpoints with no cross-handler
coupling.  Out of scope (separate future PRs): /voice/text (delegates
to /chat), the /chat family, /mode, /input/text, /dictation — those
need chat_handler + chat_msg_store + heavier surfaces and ship as
their own PRs.

## Diff

  main/CMakeLists.txt          |   2 +-
  main/debug_server.c          | 100 +-----  (4 handlers, 2 region edits, 2 URI defs, 4 register calls)
  main/debug_server_voice.h    |  29 +++  (new)
  main/debug_server_voice.c    | 137 +++++ (new)

debug_server.c:  4,023 → 3,938 LOC (−85).
Cumulative on this branch (K144 + OTA + codec + voice):
  4,520 → 3,938 = −582 LOC, **13 % shrinkage** from origin/main baseline.

## Hardware verification

build: clean (after fixing a `/chat/*` literal in the header doc-comment
        that GCC's -Werror=comment caught — `*/` ambiguity).

POST /voice/cancel  → {ok: true}
GET  /voice          → {connected: true, state: 2, state_name: "READY"}
                       (full LLM/STT text fields when populated)

story_smoke (14 steps): 14/14 PASS in 92 s.
   Step 4 "Voice ready" exercises GET /voice via the new module.
